Изменения в названиях файлови функций, для более четкого их понимания
