+++ js/main.js
@@ -1,9 +1,9 @@
 import { createPosts } from './create-array-posts.js';
 import { renderThumbnails } from './create-thumbnails.js';
-import { bigPicturePreview } from './big-picture-preview.js';
+import { ViewPictureFullSize } from './ViewPictureFullSize.js';
 
 const createdArray = createPosts();
 renderThumbnails(createdArray);
-bigPicturePreview ();
+ViewPictureFullSize ();
