+++ js/utils.js
@@ -25,4 +25,6 @@ const getRandomIdFromRangeGenerator = (min, max) => {
   };
 };
 
-export {getRandomInteger, getRandomIdFromRangeGenerator};
+const isEscapeKey = (evt) => evt.key === 'Escape';
+
+export {getRandomInteger, getRandomIdFromRangeGenerator, isEscapeKey};
